updating paths from folder restructure
