+++ base_app/src/app/app.component.spec.ts
@@ -3,11 +3,11 @@ import { AppComponent } from './app.component';
 import { AppTabNormalComponent } from './app-tab-normal/app-tab-normal.component';
 import { AppTabOverlayComponent } from './app-tab-overlay/app-tab-overlay.component';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { CommandButtonComponent } from './cmdsets/command-button';
+import { CommandButtonComponent } from './command-button/command-button';
 import { ImageOverlaysComponent } from './app-tab-overlay/image-overlays';
 import { OverlayCmdBarComponent } from './app-tab-overlay/overlay-cmd-bar';
 import { DatasetTableComponent } from './dataset-tables/dataset-table';
-import { PropDefinedTableComponent } from './dataset-tables/prop-defined-table';
+import { PropDefinedTableComponent } from './prop-defined-table/prop-defined-table';
 import { SeparatorBarComponent } from './dataset-tables/separator-bar';
 import {OverlayPageComponent} from './app-tab-overlay/overlay-page';
 import {CommonModule} from '@angular/common';
