+++ base_app/src/app/app.module.ts
@@ -6,7 +6,7 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { CmdSetComponent } from './cmdsets/cmdset.component';
 import { CommonModule } from '@angular/common';
 import { DataSetChangeService } from './services/dataset-change.service';
-import { CommandButtonComponent } from './cmdsets/command-button';
+import { CommandButtonComponent } from './command-button/command-button';
 import { ErrorPopupComponent } from './cmdsets/error-popup';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpClientModule, HttpClient } from '@angular/common/http';
@@ -24,7 +24,7 @@ import { OverlayCmdBarComponent } from './app-tab-overlay/overlay-cmd-bar';
 import { OverlayPageComponent } from './app-tab-overlay/overlay-page';
 import { DatasetTableComponent } from './dataset-tables/dataset-table';
 import { PortalModule } from '@angular/cdk/portal';
-import { PropDefinedTableComponent } from './dataset-tables/prop-defined-table';
+import { PropDefinedTableComponent } from './prop-defined-table/prop-defined-table';
 import { SectionComponent } from './section/section.component';
 import { SeparatorBarComponent } from './dataset-tables/separator-bar';
 import { CommandButtonChangeService } from './services/command-button-change.service';
